+++ R/parameters.R
@@ -254,7 +254,7 @@ cost_complexity <- function(range = c(-10, -1), trans = log10_trans()) {
 #' * `num_comp()`: The number of components in a model
 #' (e.g. PCA or PLS components).
 #'
-#' * `deg_free()`: A parameter for the degrees of freedom.
+#' * `deg_free()`: A parameter for integer degrees of freedom.
 #'
 #' * `prune_method()`: A parameter for pruming methods for MARS.
 #'
@@ -434,9 +434,9 @@ degree <- function(range = c(1, 3), trans = NULL) {
 
 #' @export
 #' @rdname para_parameters
-deg_free <- function(range = c(1, 5), trans = NULL) {
+deg_free <- function(range = c(2, 10), trans = NULL) {
   new_quant_param(
-    type = "double",
+    type = "integer",
     range = range,
     inclusive = c(TRUE, TRUE),
     trans = trans,