+++ tests/testthat/test_parameters.R
@@ -16,7 +16,7 @@ test_that('param ranges', {
   expect_equal(cost(-2:-1)$range, list(lower = -2, upper = -1))
   expect_equal(scale_factor(-2:-1)$range, list(lower = -2, upper = -1))
   expect_equal(dials::margin(c(.1, .15))$range, list(lower = .1, upper = .15))
-  expect_equal(deg_free(1:2)$range, list(lower = 1, upper = 2))
+  expect_equal(deg_free(1:2)$range, list(lower = 1L, upper = 2L))
   expect_equal(hidden_units(1:2)$range, list(lower = 1L, upper = 2L))
   expect_equal(batch_size(1:2)$range, list(lower = 1L, upper = 2L))
   expect_equal(Laplace(1:2)$range, list(lower = 1, upper = 2))