+++ NEWS.md
@@ -3,8 +3,7 @@
 ## Breaking changes
 
 * All parameter _objects_ are now parameter _functions_. For example, the 
-pre-configured object `mtry` is now a function, `mtry()`, with arguments for the `range` and the `trans`form. This provides greater flexibility in parameter
-creation, and should feel more natural.
+pre-configured object `mtry` is now a function, `mtry()`, with arguments for the `range` and the `trans`. This provides greater flexibility in parameter creation, and should feel more natural.
 
 * `deg_free()` erroneously produced real values; integers are now returned.